# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  dto.ConnectionResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      encrypted_secret_data:
        $ref: '#/definitions/types.ConnectionMetadata'
      environment_id:
        type: string
      id:
        type: string
      name:
        type: string
      provider_type:
        $ref: '#/definitions/types.SecretProvider'
      status:
        $ref: '#/definitions/types.Status'
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  dto.ListConnectionsResponse:
    properties:
      connections:
        items:
          $ref: '#/definitions/dto.ConnectionResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  errors.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      internal_error:
        type: string
      message:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      success:
        type: boolean
    type: object
...
  types.SecretProvider:
    enum:
    - flexprice
    - stripe
    - razorpay
    type: string
    x-enum-varnames:
    - SecretProviderFlexPrice
    - SecretProviderStripe
    - SecretProviderRazorpay
...
  types.Status:
    enum:
    - published
    - deleted
    - archived
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDeleted
    - StatusArchived
...
info:
  contact: {}
  description: FlexPrice API Service
  title: FlexPrice API
  version: "1.0"
paths:
...
  /integration/providers:
    get:
      consumes:
      - application/json
      description: Get all available payment providers for the current tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListConnectionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get available providers
      tags:
      - Integration
  /integration/sync/{entity_type}/{entity_id}:
    post:
      consumes:
      - application/json
      description: Sync an entity to all available payment providers for the current
        tenant
      parameters:
      - description: Entity type (e.g., customer, invoice, tax)
        in: path
        name: entity_type
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Sync entity to all available providers
      tags:
      - Integration
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key in the format *x-api-key &lt;api-key&gt;**
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"